+++ Gpufit/cuda_kernels.cuh
@@ -6,15 +6,15 @@
 void configure_model(ModelID const model_id, int & n_parameters, int & n_dimensions);
 
 extern __global__ void convert_pointer(
-    double ** pointer_to_pointer,
-    double * pointer,
+    float ** pointer_to_pointer,
+    float * pointer,
     int const n_pointers,
     int const size,
     int const * skip);
 
 extern __global__ void cuda_calculate_interim_euclidian_norms(
-    double * norms,
-    double const * vectors,
+    float * norms,
+    float const * vectors,
     int const n_points,
     int const n_fits,
     int const n_parameters,
@@ -23,31 +23,31 @@ extern __global__ void cuda_calculate_interim_euclidian_norms(
     int const n_fits_per_block);
 
 extern __global__ void cuda_complete_euclidian_norms(
-    double * norms,
+    float * norms,
     int const n_blocks_per_fit,
     int const n_fits,
     int const n_parameters,
     int const * finished);
 
 extern __global__ void cuda_sum_chi_square_subtotals(
-    double * chi_squares,
+    float * chi_squares,
     int const n_blocks_per_fit,
     int const n_fits,
     int const * finished);
 
 extern __global__ void cuda_check_fit_improvement(
     int * iteration_failed,
-    double const * chi_squares,
-    double const * prev_chi_squares,
+    float const * chi_squares,
+    float const * prev_chi_squares,
     int const n_fits,
     int const * finished);
 
 extern __global__ void cuda_calculate_chi_squares(
-    double * chi_squares,
+    float * chi_squares,
     int * states,
-    double const * data,
-    double const * values,
-    double const * weights,
+    float const * data,
+    float const * values,
+    float const * weights,
     int const n_points,
     int const n_fits,
     int const estimator_id,
@@ -57,7 +57,7 @@ extern __global__ void cuda_calculate_chi_squares(
     std::size_t const user_info_size);
 
 extern __global__ void cuda_sum_gradient_subtotals(
-    double * gradients,
+    float * gradients,
     int const n_blocks_per_fit,
     int const n_fits,
     int const n_parameters,
@@ -65,11 +65,11 @@ extern __global__ void cuda_sum_gradient_subtotals(
     int const * finished);
 
 extern __global__ void cuda_calculate_gradients(
-    double * gradients,
-    double const * data,
-    double const * values,
-    double const * derivatives,
-    double const * weights,
+    float * gradients,
+    float const * data,
+    float const * values,
+    float const * derivatives,
+    float const * weights,
     int const n_points,
     int const n_fits,
     int const n_parameters,
@@ -83,11 +83,11 @@ extern __global__ void cuda_calculate_gradients(
     std::size_t const user_info_size);
 
 extern __global__ void cuda_calculate_hessians(
-    double * hessians,
-    double const * data,
-    double const * values,
-    double const * derivatives,
-    double const * weights,
+    float * hessians,
+    float const * data,
+    float const * values,
+    float const * derivatives,
+    float const * weights,
     int const n_points,
     int const n_parameters,
     int const n_parameters_to_fit,
@@ -99,15 +99,15 @@ extern __global__ void cuda_calculate_hessians(
     std::size_t const user_info_size);
 
 __global__ void cuda_calc_scaling_vectors(
-    double * scaling_vectors,
-    double const * hessians,
+    float * scaling_vectors,
+    float const * hessians,
     int const n_parameters,
     int const * finished,
     int const n_fits_per_block);
 
 extern __global__ void cuda_init_scaled_hessians(
-    double * scaled_hessians,
-    double const * hessians,
+    float * scaled_hessians,
+    float const * hessians,
     int const n_fits,
     int const n_parameters,
     int const * finished,
@@ -115,9 +115,9 @@ extern __global__ void cuda_init_scaled_hessians(
     int const * newton_step_accepted);
 
 extern __global__ void cuda_modify_step_widths(
-    double * hessians,
-    double const * lambdas,
-    double * scaling_vectors,
+    float * hessians,
+    float const * lambdas,
+    float * scaling_vectors,
     unsigned int const n_parameters,
     int const * finished,
     int const n_fits_per_block,
@@ -125,13 +125,13 @@ extern __global__ void cuda_modify_step_widths(
     int const * newton_step_accepted);
 
 extern __global__ void cuda_calc_curve_values(
-    double const * parameters,
+    float const * parameters,
     int const n_fits,
     int const n_points,
     int const n_parameters,
     int const * finished,
-    double * values,
-    double * derivatives,
+    float * values,
+    float * derivatives,
     int const n_fits_per_block,
     int const n_blocks_per_fit,
     ModelID const model_id,
@@ -140,9 +140,9 @@ extern __global__ void cuda_calc_curve_values(
     std::size_t const user_info_size);
 
 extern __global__ void cuda_update_parameters(
-    double * parameters,
-    double * prev_parameters,
-    double const * deltas,
+    float * parameters,
+    float * prev_parameters,
+    float const * deltas,
     int const n_parameters_to_fit,
     int const * parameters_to_fit_indices,
     int const * finished,
@@ -150,10 +150,10 @@ extern __global__ void cuda_update_parameters(
 
 extern __global__ void cuda_check_for_convergence(
     int * finished,
-    double const tolerance,
+    float const tolerance,
     int * states,
-    double const * chi_squares,
-    double const * prev_chi_squares,
+    float const * chi_squares,
+    float const * prev_chi_squares,
     int const iteration,
     int const max_n_iterations,
     int const n_fits);
@@ -174,17 +174,17 @@ extern __global__ void cuda_check_all_lambdas(
     int const n_fits);
 
 extern __global__ void cuda_prepare_next_iteration(
-    double * lambdas,
-    double * chi_squares,
-    double * prev_chi_squares,
-    double * function_parameters,
-    double const * prev_parameters,
+    float * lambdas,
+    float * chi_squares,
+    float * prev_chi_squares,
+    float * function_parameters,
+    float const * prev_parameters,
     int const n_fits,
     int const n_parameters);
 
 __global__ void cuda_update_temp_derivatives(
-    double * temp_derivatives,
-    double const * derivatives,
+    float * temp_derivatives,
+    float const * derivatives,
     int const * iteration_failed,
     int const n_fits_per_block,
     int const n_blocks_per_fit,
@@ -193,9 +193,9 @@ __global__ void cuda_update_temp_derivatives(
     int const n_points);
 
 __global__ void cuda_multiply(
-    double * products,
-    double const * multiplicands,
-    double const * multipliers,
+    float * products,
+    float const * multiplicands,
+    float const * multipliers,
     int const * skip,
     int const n_vectors,
     int const vector_size,
@@ -203,9 +203,9 @@ __global__ void cuda_multiply(
     int const * not_skip_3);
 
 __global__ void cuda_multiply_matrix_vector(
-    double * products,
-    double const * matrices,
-    double const * vectors,
+    float * products,
+    float const * matrices,
+    float const * vectors,
     int const n_rows,
     int const n_cols,
     int const n_fits_per_block,
@@ -213,39 +213,39 @@ __global__ void cuda_multiply_matrix_vector(
     int const * skip);
 
 __global__ void cuda_initialize_step_bounds(
-    double * step_bounds,
-    double * scaled_parameters,
+    float * step_bounds,
+    float * scaled_parameters,
     int const * finished,
     int const n_fits,
     int const n_parameters);
 
 __global__ void cuda_adapt_step_bounds(
-    double * step_bounds,
-    double const * scaled_delta_norms,
+    float * step_bounds,
+    float const * scaled_delta_norms,
     int const * finished,
     int const n_fits);
 
 __global__ void cuda_update_step_bounds(
-    double * step_bounds,
-    double * lambdas,
-    double const * approximation_ratios,
-    double const * actual_reductions,
-    double const * directive_derivatives,
-    double const * chi_squares,
-    double const * prev_chi_squares,
-    double const * scaled_delta_norms,
+    float * step_bounds,
+    float * lambdas,
+    float const * approximation_ratios,
+    float const * actual_reductions,
+    float const * directive_derivatives,
+    float const * chi_squares,
+    float const * prev_chi_squares,
+    float const * scaled_delta_norms,
     int const * finished,
     int const n_fits);
 
 __global__ void cuda_calc_phis(
-    double * phis,
-    double * phi_derivatives,
-    double * inverted_hessians,
-    double * scaled_deltas,
-    double * scaled_delta_norms,
-    double * temp_vectors,
-    double const * scaling_vectors,
-    double const * step_bounds,
+    float * phis,
+    float * phi_derivatives,
+    float * inverted_hessians,
+    float * scaled_deltas,
+    float * scaled_delta_norms,
+    float * temp_vectors,
+    float const * scaling_vectors,
+    float const * step_bounds,
     int const n_parameters,
     int const * finished,
     int const * lambda_accepted,
@@ -253,64 +253,64 @@ __global__ void cuda_calc_phis(
     int const n_fits_per_block);
 
 __global__ void cuda_adapt_phi_derivatives(
-    double * phi_derivatives,
-    double const * step_bounds,
-    double const * scaled_delta_norms,
+    float * phi_derivatives,
+    float const * step_bounds,
+    float const * scaled_delta_norms,
     int const * finished);
 
 __global__ void cuda_check_phi(
     int * newton_step_accepted,
-    double const * phis,
-    double const * step_bounds,
+    float const * phis,
+    float const * step_bounds,
     int const * finished,
     int const n_fits);
 
 __global__ void cuda_check_abs_phi(
     int * lambda_accepted,
     int const * newton_step_accepted,
-    double const * phis,
-    double const * step_bounds,
+    float const * phis,
+    float const * step_bounds,
     int const * finished,
     int const n_fits);
 
 __global__ void cuda_init_lambda_bounds(
-    double * lambdas,
-    double * lambda_lower_bounds,
-    double * lambda_upper_bounds,
-    double * scaled_gradients,
-    double const * scaled_delta_norms,
-    double const * phis,
-    double const * phi_derivatives,
-    double const * step_bounds,
-    double const * gradients,
-    double const * scaling_vectors,
+    float * lambdas,
+    float * lambda_lower_bounds,
+    float * lambda_upper_bounds,
+    float * scaled_gradients,
+    float const * scaled_delta_norms,
+    float const * phis,
+    float const * phi_derivatives,
+    float const * step_bounds,
+    float const * gradients,
+    float const * scaling_vectors,
     int const * finished,
     int const n_fits,
     int const n_parameters,
     int const * newton_step_accepted);
 
 __global__ void cuda_update_lambdas(
-    double * lambdas,
-    double * lambda_lower_bounds,
-    double * lambda_upper_bounds,
-    double const * phis,
-    double const * phi_derivatives,
-    double const * step_bounds,
+    float * lambdas,
+    float * lambda_lower_bounds,
+    float * lambda_upper_bounds,
+    float const * phis,
+    float const * phi_derivatives,
+    float const * step_bounds,
     int const * finished,
     int const * lambda_accepted,
     int const * newton_step_accepted,
     int const n_fits);
 
 __global__ void cuda_calc_approximation_quality(
-    double * predicted_reductions,
-    double * actual_reductions,
-    double * directive_derivatives,
-    double * approximation_ratios,
-    double * derivatives_deltas,
-    double const * scaled_delta_norms,
-    double const * chi_squares,
-    double const * prev_chi_squares,
-    double const * lambdas,
+    float * predicted_reductions,
+    float * actual_reductions,
+    float * directive_derivatives,
+    float * approximation_ratios,
+    float * derivatives_deltas,
+    float const * scaled_delta_norms,
+    float const * chi_squares,
+    float const * prev_chi_squares,
+    float const * lambdas,
     int const * finished,
     int const n_fits,
     int const n_points);
