+++ Gpufit/estimators/lse.cuh
@@ -37,16 +37,16 @@
 */
 
 __device__ void calculate_chi_square_lse(
-    volatile double * chi_square,
+    volatile float * chi_square,
     int const point_index,
-    double const * data,
-    double const * value,
-    double const * weight,
+    float const * data,
+    float const * value,
+    float const * weight,
     int * state,
     char * user_info,
     std::size_t const user_info_size)
 {
-    double const deviation = value[point_index] - data[point_index];
+    float const deviation = value[point_index] - data[point_index];
 
     if (weight)
     {
@@ -101,10 +101,10 @@ __device__ void calculate_hessian_lse(
     int const point_index,
     int const parameter_index_i,
     int const parameter_index_j,
-    double const * data,
-    double const * value,
-    double const * derivative,
-    double const * weight,
+    float const * data,
+    float const * value,
+    float const * derivative,
+    float const * weight,
     char * user_info,
     std::size_t const user_info_size)
 {
@@ -159,17 +159,17 @@ __device__ void calculate_hessian_lse(
 */
 
 __device__ void calculate_gradient_lse(
-    volatile double * gradient,
+    volatile float * gradient,
     int const point_index,
     int const parameter_index,
-    double const * data,
-    double const * value,
-    double const * derivative,
-    double const * weight,
+    float const * data,
+    float const * value,
+    float const * derivative,
+    float const * weight,
     char * user_info,
     std::size_t const user_info_size)
 {
-    double const deviation = data[point_index] - value[point_index];
+    float const deviation = data[point_index] - value[point_index];
 
     if (weight)
     {
